% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bekkIRF-package.R
\docType{package}
\name{bekkIRF-package}
\alias{bekkIRF}
\alias{bekkIRF-package}
\title{bekkIRF: Impulse Response Functions for BEKK Models}
\description{
Provides a simulation-based approach to compute impulse response functions for asymmetric multivariate GARCH models.
}
\author{
\strong{Maintainer}: Ruven Zapf \email{Ruven.Zapf@uni-goettingen.de}

Authors:
\itemize{
  \item Helmut Herwartz
}

}
\keyword{internal}
