% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment_utils.R
\name{compute_psi_kurtosis}
\alias{compute_psi_kurtosis}
\title{Compute fourth-order moment matrix}
\usage{
compute_psi_kurtosis(xi)
}
\arguments{
\item{xi}{A numeric matrix of dimension \verb{T x K} containing standardized
innovations.}
}
\value{
A numeric matrix of dimension \verb{K^2 x K^2}.
}
\description{
Computes the empirical fourth-order moment matrix used in higher-moment
impulse response analysis. The matrix is defined as
\deqn{\Psi = \frac{1}{T} \sum_{t=1}^T (\xi_t \xi_t') \otimes (\xi_t \xi_t').}
}
\keyword{internal}
