% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf_compute.R
\name{irf_compute}
\alias{irf_compute}
\title{Compute an impulse response function}
\usage{
irf_compute(x)
}
\arguments{
\item{x}{A numeric input}
}
\value{
A numeric value.
}
\description{
Compute an impulse response function
}
\examples{
x <- 3
irf_compute(x)
}
