% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bekkIRF_methods.R
\name{print.bekkIRF}
\alias{print.bekkIRF}
\title{Print a BEKK IRF object}
\usage{
\method{print}{bekkIRF}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"bekkIRF"}.}

\item{...}{Additional arguments ignored.}
}
\description{
Print a BEKK IRF object
}
