% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bekkIRF_methods.R
\name{print.summary_bekkIRF}
\alias{print.summary_bekkIRF}
\title{Print a BEKK IRF summary}
\usage{
\method{print}{summary_bekkIRF}(x, ...)
}
\arguments{
\item{x}{Object returned by \code{\link[=summary.bekkIRF]{summary.bekkIRF()}}.}

\item{...}{Additional arguments ignored.}
}
\description{
Print a BEKK IRF summary
}
