% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bekkIRF_methods.R
\name{summary.bekkIRF}
\alias{summary.bekkIRF}
\title{Summarise a BEKK IRF object}
\usage{
\method{summary}{bekkIRF}(object, ...)
}
\arguments{
\item{object}{Object of class \code{"bekkIRF"}.}

\item{...}{Additional arguments ignored.}
}
\description{
Summarise a BEKK IRF object
}
